fix memory leaks and null pointers

+++ src/util.c
@@ -22,7 +22,7 @@
 #include <math.h>
 
 /* Geometry helper functions */
-float get_distance_btwn_points(int x1, int y1, int x2, int y2)
+double get_distance_btwn_points(int x1, int y1, int x2, int y2)
 {
     int dx = x1 - x2;
     int dy = y1 - y2;
