+++ src/util.h
@@ -26,7 +26,7 @@
 
 const char *generate_name(const char *filename);
 
-float get_distance_btwn_points(int x1, int y1, int x2, int y2);
+double get_distance_btwn_points(int x1, int y1, int x2, int y2);
 
 /* Iterate over lst and call `action` for each actor. */
 void each_actor(struct engine *engine, TCOD_list_t lst,
